pyproject.toml
[tool.poetry]
name = "goit-pyweb-11"
version = "0.1.0"
description = ""
authors = ["iamleeon <165953114+iamleeon@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.0"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
sqlalchemy = "^2.0.34"
alembic = "^1.13.2"
psycopg2-binary = "^2.9.9"
psycopg2 = "^2.9.9"
libgravatar = "^1.0.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
python-dotenv = "^1.0.1"
fastapi-mail = "^1.4.1"
redis = "^5.0.8"
fastapi-limiter = "^0.1.6"
pydantic-settings = "^2.5.2"
cloudinary = "^1.41.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^8.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
